Updated in order to run goals separately

+++ incubator/maven/maven-core-artifacts/maven-plugins/maven-artifact-modifier-plugin/src/main/java/no/nav/maven/plugin/artifact/modifier/mojo/LoadArtifactConfigurationMojo.java
@@ -33,6 +33,7 @@ public class LoadArtifactConfigurationMojo extends ArtifactModifierMojo {
      * @throws MojoFailureException if the plugin failes to run. Causes an "BUILD FAILURE" message
      */
 	protected final void doExecute() throws MojoExecutionException, MojoFailureException {
+		boolean foundBusConfiguration = false;
 		
 		if(ArtifactConfiguration.isConfigurationLoaded() == true) {
 			return;
@@ -46,16 +47,16 @@ public class LoadArtifactConfigurationMojo extends ArtifactModifierMojo {
 				}
 				File extractedFolder = jarArchiveManager.unArchive(a.getFile(), scriptsFolder);
 				ArtifactConfiguration.loadConfiguration(new File(extractedFolder, "moduleconfig"), environment);
+				foundBusConfiguration = true;
 				break;
 			}
 		}
-				
-		if(ArtifactConfiguration.isConfigurationLoaded() == false) {
-			getLog().warn("The deployment does not contain dependency to a wps configuration");
+		
+		if(foundBusConfiguration == false) {
+			throw new MojoFailureException(" The pom must refer to a valid \"busconfiguration\" artifact. The artifact was not found as a dependency");
 		}
 		
 		File dest = new File(targetDirectory);
-
 		if(dest.exists() && dest.isDirectory()) {
 			for( File f : dest.listFiles())
 				f.delete();
